+++ update-processor/src/main/java/ru/tbank/processor/service/TelegramClientService.java
@@ -51,6 +51,6 @@ public class TelegramClientService {
                 .chatId(chatId)
                 .text(message)
                 .build();
-        return telegramClient.execute(sendMessage); // Sending our message object to user
+        return telegramClient.execute(sendMessage);
     }
 }
