+++ update-processor/src/main/java/ru/tbank/processor/service/group/GroupChatUpdateProcessingService.java
@@ -8,16 +8,14 @@ import org.jspecify.annotations.NullMarked;
 import org.springframework.stereotype.Service;
 import org.telegram.telegrambots.meta.api.objects.Update;
 import org.telegram.telegrambots.meta.api.objects.message.Message;
-import ru.tbank.common.entity.dto.DeletedTextMessageDTO;
+import ru.tbank.common.entity.ChatModerationSettings;
 import ru.tbank.common.entity.text.TextModerationSettings;
 import ru.tbank.common.entity.text.TextProcessingResult;
-import ru.tbank.processor.service.UpdateProcessingService;
-import ru.tbank.processor.service.persistence.AppUserService;
 import ru.tbank.processor.service.ChatModerationSettingsService;
-import ru.tbank.processor.service.persistence.DeletedTextMessageService;
-import ru.tbank.processor.service.persistence.GroupChatService;
 import ru.tbank.processor.service.TelegramClientService;
+import ru.tbank.processor.service.UpdateProcessingService;
 import ru.tbank.processor.service.group.filter.text.TextFilter;
+import ru.tbank.processor.service.persistence.GroupChatService;
 
 import java.util.Collections;
 import java.util.List;
@@ -31,8 +29,7 @@ public class GroupChatUpdateProcessingService implements UpdateProcessingService
 
     private final TelegramClientService telegramClientService;
     private final ChatModerationSettingsService chatModerationSettingsService;
-    private final DeletedTextMessageService deletedTextMessageService;
-    private final AppUserService appUserService;
+    private final DeletedMessageReportService reportService;
     private final GroupChatService groupChatService;
     private final List<TextFilter> textFilters;
 
@@ -46,18 +43,11 @@ public class GroupChatUpdateProcessingService implements UpdateProcessingService
         // TODO: логику необходимо поменять так, чтобы конфиг создавался только при добавлении бота
         //  в чат пользователем с ролью owner, иначе бот самостоятельно удаляется из чата
         var message = update.getMessage();
-        var config = chatModerationSettingsService.getChatConfig(message.getChatId());
+        var chatSettings = getSettingsForChatByMessage(message);
+        groupChatService.findById(message.getChatId())
+                .orElseGet(() -> groupChatService.save(message.getChatId(), message.getChat().getTitle()));
 
-        if (Objects.isNull(config)) {
-            log.warn("Config is null! Creating the default config...");
-            chatModerationSettingsService.createChatConfig(message.getChatId(), message.getChat().getTitle());
-            config = chatModerationSettingsService.getChatConfig(message.getChatId());
-            groupChatService.save(message.getChatId(), message.getChat().getTitle());
-        }
-
-        log.debug("Config for this chat: {}", config);
-
-        processTextMessage(message, config.getTextModerationSettings());
+        processTextMessage(message, chatSettings.getTextModerationSettings());
     }
 
     private void processTextMessage(Message message, TextModerationSettings settings) {
@@ -69,13 +59,24 @@ public class GroupChatUpdateProcessingService implements UpdateProcessingService
                 .findFirst()
                 .ifPresent(result -> {
                     telegramClientService.deleteMessage(message);
-                    var deletedTextMessage = DeletedTextMessageDTO.buildDto(message, result);
-                    var user = message.getFrom();
-                    appUserService.saveRegularUser(user.getId(), user.getFirstName(),
-                            user.getLastName(), user.getUserName());
-                    deletedTextMessageService.save(deletedTextMessage);
+                    reportService.saveReport(message, result);
                 });
 
         log.debug("Processed message id={}", message.getMessageId());
+    }
+
+    private ChatModerationSettings getSettingsForChatByMessage(Message message) {
+        var config = chatModerationSettingsService.getChatConfig(message.getChatId());
+
+        if (Objects.isNull(config)) {
+            log.warn("Config is null! Creating the default config...");
+            chatModerationSettingsService.createChatConfig(message.getChatId(), message.getChat().getTitle());
+            config = chatModerationSettingsService.getChatConfig(message.getChatId());
+            groupChatService.save(message.getChatId(), message.getChat().getTitle());
+        }
+
+        log.debug("Config for this chat: {}", config);
+
+        return config;
     }
 }
