+++ update-processor/src/main/java/ru/tbank/processor/service/persistence/GroupChatService.java
@@ -3,6 +3,7 @@ package ru.tbank.processor.service.persistence;
 import lombok.RequiredArgsConstructor;
 import org.jooq.DSLContext;
 import org.springframework.stereotype.Service;
+import ru.tbank.processor.excpetion.EntityNotFoundException;
 import ru.tbank.processor.generated.tables.GroupChat;
 import ru.tbank.processor.generated.tables.records.GroupChatRecord;
 
@@ -15,11 +16,11 @@ public class GroupChatService {
     private final GroupChat table = GroupChat.GROUP_CHAT;
     private final DSLContext dslContext;
 
-    public void save(Long chatId, String name) {
+    public GroupChatRecord save(Long chatId, String name) {
         var storedUser = findById(chatId);
 
         if (storedUser.isPresent()) {
-            return;
+            return storedUser.get();
         }
 
         var newRecord = dslContext.newRecord(table);
@@ -27,6 +28,9 @@ public class GroupChatService {
         newRecord.setId(chatId);
         newRecord.setName(name);
         newRecord.store();
+
+        return findById(chatId).orElseThrow(
+                () -> new EntityNotFoundException("User with id=%d not found".formatted(chatId)));
     }
 
     public Optional<GroupChatRecord> findById(Long chatId) {
